+++ src/Commands/NotifierStorageBackupCommand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Devuni\Notifier\Commands;
 
 use Devuni\Notifier\Concerns\ChecksNotifierEnvironment;
+use Devuni\Notifier\Concerns\DisplayHelper;
 use Devuni\Notifier\Services\NotifierConfigService;
 use Devuni\Notifier\Services\NotifierStorageService;
 use Illuminate\Console\Command;
@@ -12,6 +13,7 @@ use Illuminate\Console\Command;
 final class NotifierStorageBackupCommand extends Command
 {
     use ChecksNotifierEnvironment;
+    use DisplayHelper;
 
     protected $signature = 'notifier:storage-backup';
 
@@ -19,6 +21,8 @@ final class NotifierStorageBackupCommand extends Command
 
     public function handle(NotifierConfigService $configService, NotifierStorageService $storageService): int
     {
+        $this->displayNotifierHeader('Storage Backup');
+
         if ($this->checkMissingVariables($configService) === self::FAILURE) {
             return self::FAILURE;
         }